Blocked capture log,  using and log close
We are blocking this because we dont want them to run. When we use this simulation code in the midde of other do-file to get a simulated data in between, the "log close" specially closes the log file opened in that another do-file, and at last gives error like "No log file open".